+++ documentation/changelog.md
@@ -1,6 +1,8 @@
 # Change Log
 <pre>
-v5.0.0 Updated Facebook API to 4.7 in iOS, that supports iOS9 [TIMOB-19383]
+v5.1.0 Updated Facebook API to 4.9.1 in iOS, that supports the messenger API. [MOD-2180]
+
+v5.0.0 Updated Facebook API to 4.7.0 in iOS, that supports iOS9 [TIMOB-19383]
 	   Updated Facebook API to 4.7.0 in Android [TIMOB-19577]
 
 v4.0.5 fixed presentSendRequestDialog with to and title params [MOD-2126]